requirements.txt
pytesseract==0.3.10
pdf2image==1.16.0
Pillow==9.5.0
flask==2.3.2  
goggle==1.0.0